Minimal code to make Content API show a page
Note that this code disregard the lifetime of many objects, next
patches will fix that. The network bits of the code were heavily
inspired on how content_shell does things, so we need more
investigation there too.

To compile it, you need to add the line

	'../cameo/cameo.gyp:*',

to the target "All"'s dependencies list. This will include cameo into
the regular GYP build of chromium. You will also need to enable Aura
(but not Aura Shell), so when calling gyp_chromium, use:

    $ GYP_DEFINES=use_aura=1 build/gyp_chromium

Aura (not the Aura Shell) provides internally a way to create windows
without having to deal with Linux and Windows specific code. There are
plans to move Chrome in those platforms to use Aura UI. That code is
also shared with ChromeOS.