Support using favicon as app icon
This patch is to make native app window use favicon as its window icon.
Both ICO and PNG file are supported for both Windows and Gtk.

The following example HTML code shows how to use favicon meta tag:

<html>
<head>
<link rel="icon" href="16x16.png">
</head>
...
</html>

As a result, the icon 16x16.png will be used as the icon showed on both
window title bar and task bar.

Known issues:
o [GTK] App icon of GtkWindow can not be updated to the expected favicon. The
  reason is, gtk_window_set_icon has no any effect after gtk_window_show
  calling. The same issue also exists in Google Chrome.

o [Windows 7] The icon won't show on title bar if using custom frame. It
  could be addressed by forcing NativeWidgetWin to use native frame.

BUG=https://github.com/otcshare/cameo/issues/72
TEST=CameoRuntimeTest.FaviconTest_ICO
TEST=CameoRuntimeTest.FaviconTest_PNG