Add Cameo extension subsystem
The Cameo extension subsystem is a set of classes that can be used by
Browser and Renderer processes to run native code that respects a
certain interface.

The interface is minimal and exposes message passing between the
extension and the JS code running in the Renderer process. Extensions
currently are run in the Browser process, either on the FILE or UI
threads.

A more detailed design document is available at
https://github.com/otcshare/cameo/wiki/Message-passing-extensions

BUG=https://github.com/otcshare/cameo/issues/65