[TestUtils] Fix that browsertest will hang sometimes at trybots
The hung process is the render process of one of the test cases.
The stack trace shows that it hangs on a dead lock of tc_malloc.

The root cause of the deadlock is not found yet.
A guess will be that some thread get the lock but not existed in
the forked process.

The solution is referencing upstream issue:
https://code.google.com/p/chromium/issues/detail?id=177218

BUG=https://github.com/otcshare/build-infrastructure/issues/107