Run extensions in their own threads
Changes our extension system so extensions are run in their own threads. This is
an improvement since extensions blocking for any reason won't impact the rest of
the system.

CameoExtension interface was changed to be a factory of Contexts, that can keep
state and handle the messages from one WebContents. The per-Browser object is
now CameoExtensionService, and it deals with extension registration and plugging
the system into new WebContents.

For now, the filtering in IO-thread was removed for now for two reasons:

- The IDMap to match ids to correct RenderWidgetHost (and then WebContents was
  supposed to run on the UI process-only).

- render_view_id and even the actual render process that a WebContents run might
  change during the runtime. We need either another ID mapping to use or track
  the changes.

That means that messages go through UI thread before reaching the extension
thread. We'll revisit this choice to support out of process extensions and
dedicated channels.

Last but not least, an automated browsertest for extensions was added.