Use base::Value instead of std::string on the PostMessage
By using base::Value, we can let the user pass a flexible list
of parameters to PostMessage that can include different data types
(including objects) that will be serialized for us by
V8ValueConverter. This is a enabler for using the data serializers
generated by IDLs. The implementation keeps backwards compatibility
with the existing implementation, so passing a string is still supported.