Workaround crash on exit (issue #178).
Waiting on the browser thread is frowned upon since it might cause jank
(there's no way to know how much time the extension thread might be
blocked for).

However, it's not safe to destroy the extension context from a thread
other than the extension thread.  We temporarily turn off this
restriction, wait for the extension thread to be properly destroyed,
and enable the restriction again.