Add the foundations for writing internal APIs
Internal APIs will be shipped with Crosswalk and have access to a
more complete API. An extension declared using an IDL will have
data serializers generated automatically that can be used for
easily parsing and validating parameters.

The extension should derive from XWalkInternalExtension and add its
IDL to the buildsystem.