Use DidCreateScriptContext hook to install extension code in page
The hook we used before was being called twice for each context.
DidCreateScriptContext is guaranteed to be called only once.

We prefer to add a more general hook for XWalkContentRendererClient to
call instead of letting it use InstallJavaScriptAPIs() directly so
that changes to the way extension inject javascript doesn't affect
non-extension code.