Evaluate JS API code for extensions in a function context
Crosswalk Extensions provide a JavaScript source code to be executed in
the page context and expose the desired functionality. We previously
were simply evaluating this code.

This scheme is changed so that code will be executed inside a function
context, and properties that the extension wants to expose should be
added to the exports object. One positive side effect is that now one
does not need to use IIFE to define variables not exposed to the page,
in practice saving one indentation level.

To make this work, we restrict an extension named foo.bar to add
properties only to the object foo.bar.

Executing in a specific context will allow us to expose more
functionality to that extension code without exposing that functionality
to the JS application code.