Add "missing" var to the wrapped code
Use var when bootstrapping the namespace for an extension. For example,
use

 var xwalk = xwalk || {};
 xwalk.experimental = xwalk.experimental || {}; // ...

instead of

 xwalk = xwalk || {};
 xwalk.experimental = xwalk.experimental || {}; // ...

Given that they are executed in the global context, I wouldn't expect a
difference. But I observed while trying other external extensions that
using var would make the whole namespace tree accessible. This is
sometimes useful for extensions that are part of a group, so may need to
add things to their parent namespace.

I couldn't investigate further, but given that using var in that context
is fine practice, I decided to add this.