Make XWalkExtensionRunner an interface
The concrete implementation is now called
XWalkExtensionThreadedRunner, since it runs the extension context in a
thread. The separation will be useful to provide the
user (XWalkExtensionWebContentsHandler) different runners that run the
context different ways: in a thread, or in other process.

A change was made to the interface, it takes a Client instead of
taking the IPC::Sender/render_view_id to send the message. This has
the good side effect of making the extension runner not care about
potential changes in the render_view_id, or even to be used in cases
where the render_view_id do not apply (e.g. when running in the other
process).

The files were added to extensions/common/ since they'll be used by
both browser and extension processes.