Add support for JS to send synchronous messages to extension
Support a new operation for JS API for extensions to send synchronous
messages to the extension, and get a reply back. It can be used on the
JS by calling

    var reply = extension.internal.sendSyncMessage("message");

and handled at extension level by the new HandleSyncMessage(). The
choice of the term Send is to differentiate from the term Post which
is associated with asynchronous call.

In our current public API (in C) we expose a new callback in the
context struct for handling sync messages. The handler implemented
should call a new exposed function to set the reply. This approach
avoids having to deal with memory ownership of the reply.

I didn't used the versioning on C API for two reasons: we don't care
about versions of the external extension API until we do a proper
release, and this C interface will suffer a non-compatible change soon
-- after the feedback we got developing until now.

This commit adds tests for the sync messages. This feature is disabled
on Windows, follow up added in the issue
https://github.com/otcshare/crosswalk/issues/268.

NOTE: We'll make this clearer in the APIs and documentation but use of
synchronous calls to extensions is discouraged.