rpm: Add an initial packaging/ directory for Tizen 2.1 mobile.
This is an initial version with the same contents as [1] for packaging
Crosswalk on Tizen 2.1 mobile.

As described in [1], it uses the Aura backend without any GTK+2 dependency.
This new directory should probably have a repository of its own, but I am
maintaining it here for now while that gets sorted out.

Besides the TODO's in the spec file, the most outstanding item that needs to
be solved is that xwalk crashes when being launched due to some graphics
issue.

[1] http://linux.intel.com/pipermail/cameo-dev/2013-July/000640.html