rpm: Stop crashing at launch when using Aura.
This is what was going on when Crosswalk crashed:

 1. The X11-GL GPU code in Chromium (gl_implementation_x11.cc) was looking
    for a GL implementation before a GLES one by default. Since the PowerVR
    libraries used in Tizen 2.1 mobile only implement GLES, symbols were not
    found.

 2. Chromium then tries the GLES implementation by loading, among other
    shared libraries, libGLESv2.so.2. The PowerVR libraries, on the other
    hand, ship libGLESv2.so.1, so we need to make the code look for the
    "right" file.

Item 1 is currently solved by simply passing --use-gl=egl in the xwalk
wrapper we install in /usr/bin. It could be possible to change the lookup
order in gl_implementation_x11.cc, but the less we patch the source, the
better, and the effect would be exactly the same.

It must be noted that this approach does not scale across Tizen platforms:
IVI uses Mesa instead of the PowerVR libraries, for example, and Tizen PC
can probably be considered a normal desktop environment.