Move the Sync messages to use base::Value instead of std::string
This is basically the same change we did for the async messages. It will
save a few copies (currently we copy the string is some places, specially
when posting it to another thread). But the main benefit is it will be
easier to work with the serializers created from IDL files.