packaging: Change the order we use to look for GL implementations.
We need to use EGL on Tizen Mobile devices and OSMesa in the emulator (the
EGL implementation in 2.1 does not work as expected in the emulator and
crashes Crosswalk).

The easiest way for this to work is to look for the GL implementations
available in a different order: if we ship libosmesa.so in the emulator RPM
package, we can look for OSMesa first and, if it is not available, we fall
ack to EGL (which means a real device without libosmesa.so in the system)
and nly then look for the Desktop GL implementation (Tizen PC, maybe).

Once we move to a Tizen version whose emulator image works correctly with
GL, we can drop this patch.

Related to issue #272.