Fix race condition acessing context_ in XWalkExtensionThreadedRunner
This patch makes sure that "context_" is only used inside the
extension thread. This is good by itself -- accessing data from only
one thread, and also fixes a concrete flakiness we had in
context_destruction.cc test.

When we dispatched creation of a new context, and later (before
creation was done) we dispatched a "post message", the post would use
an outdated context_ (NULL). Even with the ordering of the tasks,
after creating, we would try to call a method of context_.