When navigating away, refresh the extension contexts
We should delete old extension contexts and create new ones after
navigating away from a page. The JS context is already recreated but
the native side was being kept.

This commit creates a new message DidCreateScriptContext to notify our
system at browser process that we will re-install the JS APIs in a new
context. For now support is limited to the main frame, fix for that
will come in a follow up patch.