Create constructor for XWalkExternalExtension that takes NativeLibrary
This is preparation to support both current and the new C APIs. The
plan is to land support for the new API and gradually upgrade the
external extensions we have, and finally remove the old one.

This will be useful because we'll load the NativeLibrary first to
identify whether we are dealing with the current or new C API, before
creating the class.