Make XWalkExtensionService identify libraries that are external extensions
Now XWalkExtensionService identify if the library contain the right
entry point, and if that's the case, passes it to
XWalkExternalExtension. This will be useful when we have the new C API
that contains a different entry point, because we'll be able to choose
the right class to pass the library.

As a bonus, we get a nice message when we can't load the library or
can't find a suitable entry point.