Add new Crosswalk Extensions API
This is the first step to change our C API. We add the new extension,
support loading both kinds, move our external extensions, and then
remove the support for the old. This strategy allow us to update
things gradually instead of forcing all the updates happen in the same
time.

The first API was done to get us started but some problems were
identified during its usage for developing a series of different
extensions. Note that we expected a major overhaul after the first
round development.

== INTERFACE ==

The problems identified:

(1) Our API exposed internal data unnecessarily;
(2) We were open to have per-context function for posting message and
    callback for handle message, although in practice we never used
    this.
(3) The name "Context" was confusing, the name "Instance" evokes the
    Class/Instance metaphor that makes more sense;
(4) Because the API expected extension to allocate structs, we had to care
    about their de-allocation;
(5) It was not straightforward to add new internal capabilities without
    mixing them with the rest of API. (handle_sync_message)

And how they were solved:

(1) Get rid of the "void* internal_data"; use specific non-pointer
    identifier for Extensions and Instances (former Contexts).
(2) Now the PostMessage function and HandleMessageCallback are the
    same for all instances. We now register the callback passing
    the XW_Extension.
(3) "Context" was renamed to "Instance" in the API. This new name
    reinforces the similarity with the Class/Instance concept from OO.
(4) No more structs from extension to represent Extension and Context,
    new identifiers XW_Extension and XW_Instance are used. In
    particular the extension is now informed that a new instance was
    created instead of ask to create one.
(5) The usage of GetInterface() allowed us to move sync messaging to
    a different header in a clean way.

This new API was very influenced by PPAPI, with one important
difference: extensions are not expected to implement named
interfaces (i.e. provide their own GetInterface). Instead of that, the
approach is to rely on Register*() functions from the runtime
interfaces to provide callbacks.

This approach also made less necessary the convention to prefix
functions that should be called by extension or by the runtime.

== IMPLEMENTATION ==

XWalkExternalAdapter was added to map C calls (with XW_Extension /
XW_Instance) and to the corresponding C++ objects (of types
XWalkExternalExtension and XWalkExternalContext). It is important to
note that changes were restricted to inside XWalkExternalExtension
implementation.

As part of the re-work, a significant amount of warning messages were
added to XWalkExternalExtension to identify the reasons why a certain
extension was not loaded.