Create a new layout for NativeAppWindow in Views
The TopViewLayout manages up to two children: the content view and an
optional top view. If there's a top view, it will be layouted with its
preferred height and fill the width with the window width. The remaining
space will be given to the content view.

This is a preparation for landing support for Indicator (the bar on
top with date, battery info and other data) in Tizen Mobile. Also can
be useful for other platforms. Since the layout works correctly
without setting a top view, we use it for all cases.

TEST=Added a new unit test for layout behavior; tested old behavior of
     window manually resizing it and inspecting the content; tested creating
     a top view that changes its preferred size after a timeout and the
     window was correctly repainted.