Add TizenPlug and TizenIndicator
Indicator is used on tizen mobile as a status bar, showing mobile's clock,
wi-fi, bluetooth infos and so. It's also used as a shortcut to homescreen
and settings (when scrolled down).

In this initial implementation the TizenPlug gets an image from which other
process created, in this case tizen mobile. This image is provided by a
socket and we watch a file descriptor to get this image updated.
The TizenIndicator shows this image.

The platform expects that we send touch events when the user interacts
with the indicator we painted. This is not supported yet

To enable it you should use -Dtizen_mobile=1 (only for tizen mobile builds).