[Android] Fix the issue of mixed context and activity
This is to resolve the two contexts needed by runtime side
and activity is used by runtime core.
1. Keep the mixed context inside of runtime. One context is
for application and one is for package. They're different
because runtime library is shared so the context for runtime
library is different from the context for web app APK.
2. Pass activity from runtime client to runtime. This is needed
for fullscreen support like video playback.