Add XWalkExtensionModule to expose 'extension' object
An extension module takes an extension name paired with its JS API code,
and runs it. The extension module lifetime is coupled with the module
system, so when a new v8::Context is created for a frame, we create an
extension module for each extension. We invalidate and destroy them when
the context is released by the frame.

The 'extension' object exposed when we run the JS API code is now
implemented directly in C++ instead of relying on old xwalk.*
namespace. With this, we finally isolated message passing for each
extension. We still have xwalk.* globals for the sake of internal
extension helpers, but they'll move to an appropriated module soon.