Rename few functions on the Internal Extension JS API
Rename the odd setMessageListenerInternal which is not exactly what the function
does. Also rename message_listeners to callback_listeners to avoid confusion when
Event Listeners gets implemented.

A namespace was also added, so we now access the Internal Extensions API through the
extension._internal object, making the names of the functions shorter,
because we don't need the "Internal" suffix anymore.