Add support for requireNative() in JS API code
This adds the basic support for JS API code of the extensions to use
utilities implemented in C++ on demand. When calling
requireNative("name"), the module system will lookup the right
XWalkNativeModule, call its NewInstance() method and return that object.

This is the basic building block for adding functionality that needs to
be implemented in the Renderer Process.