Add unittest to ensure threaded runner run functions in right threads
It is up to the threaded runner to ensure that: context functions are
called in the extension thread, the functions handled by the client
are NOT called in the extension thread and the client is not called
after the object is deleted.

The last two are not passing, next commit will fix them.