Add a MessageFilter for the Extension messages
Handling Extensions filters in the IO Thread can save a hop into
the UI Thread. Messages are now dispatched directly to the Extension
Thread.

Each WebContentsHandler creates its own filter that filters messages
based on the routing ID of the WebContents. The filter gets removed
by the Channel automatically on shutdown or explicitly when the
runtime is removed. In any case, WebContentsHandler will always outlive
the filter.