Make sure threaded runner calls the client in the right thread
We were calling the client in the extension thread, which is not what
XWalkExtensionThreadedRunner is supposed to do.

The object now takes a task runner in the constructor, that will be used
to post the calls to client. For now Crosswalk passes the task runner of
the thread in which we created the extension threaded runner.

The failing unittests from last commit now pass.