Do not block the IO Thread when waiting for a sync reply
Synchronous messages are not exactly synchronous from the browser
process perspective. By handling them with IPC_MESSAGE_HANDLER_DELAY_REPLY,
we have access to specially crafted message that can be used to
send the reply to renderer process. This reply can be sent at any
moment but if not, the callee on the renderer process will block
indefinitely.