Remove references to `selinux' from the build system.
This is basically the same thing as Chromium r200838: the SELinux code is
unmaintained, and support for it was removed from Chromium.

We weren't using SELinux ourselves, so removing the code does not make us
lose any features and makes us more future-proof against future rebases of
our Chromium tree.