Make our v8 usage friendlier to upcoming changes
In V8 API, there was a big change (already in trunk but still not in our
branches) on the way v8::Persistents can be created / used. Before the
change, v8::Persistent could act as a v8::Handle, so we could "use" a
value directly. With the new behaviour, we need to unwrap them for
using.

This patch adds the code to unwrap them, reducing the amount of change
needed to run Crosswalk in trunk.