Make XWalkViewsDelegate inherit from TestViewsDelegate
This patch brings the special case for OnBeforeWidgetInit() function to
our own class which is the only thing we used from
DesktopTestViewsDelegate. This is a first step to inherit directly from
ViewsDelegate instead of relying in testing infrastructure.

Also removed g_views_delegate since views::ViewsDelegate already provide
a place to store it's application-wide instance.