Add TizenPlugMessageWriter and TizenSystemIndicator event support.
This class is used to build and write messages using the protocol
implemented in EFL 1.7 on the Tizen Mobile 2.1 platform.

Messages contain a header and a payload. The instructions are used
to build the new header based on a previous header. This allows the
system to save bytes when a certain header field wasn't updated.

This utility class is used to construct the next message based on the
previous one and the new instructions.

Now TizenSystemIndicator and TizenSystemIndicatorWatcher support
click and touch events.