Add a JavaScript object lifecycle tracker
This helper works like this: create an instance by calling
lifecycleTracker() and set the "destructor" property of tracker
to the callback you want invoked when the tracker is about to die.

This is useful to track the lifecycle of objects that have an instance
on the native side. If they get collected by the GC, we need to somehow
notify the extension.

The usage in this case is simple: just assign the tracker as a property
of the object you want to track. The callback of the tracker will be
called when no reference to the object containing the tracker is held.