Disable DragDrop support from DesktopRootWindowHostXWalk
That class is a copy of an older version of DesktopRootWindowHostX11,
and uses its helper object. It happens that with the new version,
DesktopDragDropClientAuraX11 is tied to the type of
DesktopRootWindowHostX11, so we break.

To move forward with this, we should try: (a) upstream whatever changes
we think are relevant for X11 general version to avoid having our own in
the future  (b) try to make DesktopRootWindowHostXWalk a subclass of
X11, it's tricky but with small changes we can do it.

This patch fixes the build, but still blank screen when loading a page
using Aura build on my Linux Desktop.