Use future-proof includes when possible
Chromium usually move include files in two steps, first create one in
the new location and make the old point to the new; in a second step
later, remove the old one.

When available, we should use the new location even if the older one
works, so that future updates are less painful, and people using
Crosswalk with trunk have less changes to make.

This patch changes:

- string_util.h is now in base/strings/
- stringprintf.h is now in base/strings/
- message_loop.h is now in base/message_loop/message_loop.h

This patch shouldn't change behavior of Crosswalk.