[Android] IntentFilter support in app wrapper.
Intent messaging is a facility for late run-time binding
between components in the same or different applications,
which is passed when initiating the Activity and Service,
or from any of the broadcast methods.
IntentFilter is used to filter the intents.

Take remote debug as example:
adb shell am broadcast -a org.xwalk.intent -e remotedebugging true