[Android] Implement the runtime API - getVersion
'getVersion' is important for it'll be used to support backward
compatibility and upgrade. Now two parts are generated automatically.

For runtime library part, we use the native macro 'XWALK_VERSION' which
is defined by gyp. For runtime client part, a python script is used to generate
a Version Java file with the version from build system.