Make new C API code more robust, and test it
Instead of CHECK() whether XW_Instances and XW_Extensions coming from
the external extensions are correct, we now ignore calls to invalid
instances and extensions and loudly emit a warning with detailed
information.

Note that if multiple extensions or instances are running in a process,
they still can "guess" a valid (but incorrect) id.

This problem was identified when testing the new API with t-e-c, where
some extensions send messages to older (already destroyed) instances.