Support install application on Tizen
When running in Tizen mobile OS, it'll call install_into_pkginfo_db.py to install package information into Tizen system database, create icon on home screen.

It support launch an application by it ID which is come from progame's
base name. Since in Tizen, application is launched by its symbolic link
to the xwalk binary, and the link name is application's ID, it's the
only way to get it.

In Tizen, the DATA_PATH should be fixed, since the general user name is
"app", but when install, it's super root, and the default DATA_PATH is
$HOME/.config/xwalk, so the real DATA_PATH will be changed when switch
between these two users, and runtime can not launch any installed
applications because of no record in database.

Add the default command starting parameters for Tizen, e.g.
"--fullscreen", "--use-gl=xxx", "--ignore-gpu-blacklist",
"--allow-file-access-from-files", otherwise, the application can not be
launched successfully.