Split .gclient-xwalk generation and gclient sync into two scripts.
Besides better separating two different activities, it helps a lot in the
case of automatic builds: if we are actually interested in building for a
different platform (Tizen, for example), we do not need to run hooks (such
as gyp_xwalk or download NaCl toolchains), in fact doing so generates
additional files and requires additional dependencies that are not needed at
all.

There should be no visible difference to those who just run `gclient sync',
but if one runs fetch_deps.py it will only sync an existing .gclient-xwalk,
which should have been previously generated by generate_gclient-xwalk.py.