Use CancellationFlag instead of Lock in PostHelper
XWalkExtensionThreadedRunner have a helper object that runs jobs in
another thread. When threaded runner goes down, we want to
invalidate (cancel) that helper object.

This commit change from using a Lock and changing the runner_ value
inside the helper to simply use CancellationFlag (which boils down to an
atomic operated value).