[Android] Remove some of the duplicated code for handling Asset URLs.
The code paths in RuntimeURLRequestContextGetter::GetURLRequestConext() can
be greatly simplified if we stop using a separate job factory and a
single-item vector and just add a ProtocolInterceptJobFactory when on
Android for the asset:// protocol handler.

While here, we can also remove XWalkUrlRequestJobFactory, whose special
functionality is not being used anywhere (if it ever was).