Create a ModuleSystem even for blank pages
We assumed that a v8::Context associated with a frame that were pointing
to "about:blank" was never reused when visiting other URL. This
assumption is incorrect, there are situations (like the cases triggered
by the bugs linked) that the v8::Context is reused to the page loaded.

Since the frame URL can change from "about:blank" to something else
between the calls DidCreateScriptContext() and
WillReleaseScriptContext(), we can't rely on that information. The crash
was caused by assuming that if the URL wasn't blank at
WillReleaseScriptContext() time, we had a module.

The intent of not creating module system for blank pages was to avoid
unnecessary work. This will be accomplished in the future by other means
like avoiding to load the JavaScript code that's not used.

Fixes #626 and #602.