[Extensions] Create a ModuleSystem even for blank pages
Looks like they can be changed to something else and the contexts are
reused, so we don't get the DidCreateScriptContext again and end up
by free()'ing an invalid memory space.

In the feature, when lazy loading is implemented, we won't be affected
by wasting time creating ModuleSystems for pages that stay blank forever.

Fixes #626 and #602.