Change ContextDestruction test to be less picky about titles
For the test move forward, we depended on the exact sequence of
instances being created. We can cover the same behavior (which is
instances will be created and not reused), without having to indicate
this in the title.

This allows us to be be lax about the exact number of instances created
now, until we add support for loading instances on demand ("lazy loading
of instances").