[Android] Add navigation handler for handling platform-specific schemes.
Some platform-specific schemes are needed to be handled, eg, "wtai://", we
provide default navigation handler in XWalkView to support such schemes.

"wtai://" is a protocol scheme that can initiate a dialog with the mobile phone
user whether to place a call to that number, send an SMS message, or store the
number in the address list of the mobile phone.

See: http://www.w3.org/wiki/UriSchemes/wtai

In this patch we support placing a call. And other functions will be added later.

BUG=https://github.com/crosswalk-project/crosswalk/issues/380