DEPS.xwalk: Stop using SSH for the crosswalk repositories URLs.
Now that the crosswalk repositories are public, we can use direct https URLs
to clone them instead of the previous ones that used SSH.

The overhead of using SSH may be negligible in terms of speed, but doing so
allows truly anonymous checkouts: people may not have a GitHub account (or
have their SSH keys in the machine they're using at the moment).

Note that git:// would also work, but can cause problems for people behind
firewalls.