[Android] Cancel library-not-found dialog when library installed
For xwalk apps on Android, they will try to find runtime lib at
startup time. If runtime lib not found, an alert dialog will be
prompted for navigating user to download it from play store.

When the app is brought back to the front, it will check
runtime lib's existense. If runtime lib found, the alert dialog
should be automatically canceled.

BUG=https://github.com/crosswalk-project/crosswalk/issues/638