[Build] Generate `xwalk_version' from the VERSION file.
One less place that defines the current version, we're on our way to
defining it in only one file.

Borrow Chromium's chrome/tools/build/version.py utility to read the value in
VERSION and get an appropriate version number for it.

Some places in xwalk.gyp depend on the version but are not automatically
regenerated/rebuilt when VERSION changes, but that was also the case before,
and fixing that isn't trivial, so let's defer that part.