Add main document launch support
Main document is an invisble document no matter whether it is a real file or a
generated one. Working as the main entry of an application, the main document
must be loaded before other pages.(see
http://anssiko.github.io/runtime/app-lifecycle.html)

So the Runtime::Create() is modified to create a Runtime instance without attach
to any app window, and the WebContents' initial size is also set to (0,0).
Runtime::CreateWithDefaultWindow() now works like previous Runtime::Create().
Currently we support application both run from main document and local path,
however the local path support may be deprecated in the future.