[WiDi API] [RE] First commit for Presentation API implementation on Android
RECOMMIT it to add backward compatibility support for Android 4.x device.

NOTE: Since the current implementation relies on the DisplayManager
class introduced from Android 4.2 (API Level 17), as a result, the
Presentation API is only available on Android 4.2+, but won't break
running on the older version.

This commit is the first patch to implement Presentation API spec on
Android. It is intended to bring wireless display support into Web. The
new API is placed under 'navigator.presentation' namespace, and the
'displayavailable' property and its change event is implemented in this
commit by use of DisplayManager in Android SDK.

SPEC=http://otcshare.github.io/presentation-spec/index.html
BUG=https://github.com/crosswalk-project/crosswalk/issues/577