[WiDi API] First commit for Presentation API implementation on Android
This commit is the first patch to implement Presentation API spec on
Android. It is intended to bring wireless display support into Web. The
new API is placed under 'navigator.presentation' namespace, and the
'displayavailable' property and its change event is implemented in this
commit by use of DisplayManager in Android SDK.

SPEC=http://otcshare.github.io/presentation-spec/index.html
BUG=https://github.com/crosswalk-project/crosswalk/issues/577