[Tools] Add bump-canary-version.sh.
This script is used to automatically bump Crosswalk's build version (ie. the
"3" in version "1.2.3.4"), so that the automated Canary generation machinery
can used it to bump versions and commit when all canaries build and pass all
tests.

Using a simple, POSIX shell script with Perl one-liners is much quicker than
writing a full-fledged Python script just to perform some glorified sed's,
and I hope the Perl is easy enough to understand.