Clean up in use of DesktopRootWindowHostXWalk
DesktopRootWindowHostXWalk serves as our "desktop environment"
per scree, which is somewhat similar to the X11 window manager
concept.

The specialized version of DesktopRootWindowHost (*XWalk) in
our situation, is created as part of DesktopNativeWidgetAura
initialization, but requires build system changes for doing so.

This patch cleans up part of the code, making it very clear
why we manually creates DesktopRootWindowHostXWalk and why it
works. It also removes the ::Create method which is called
by DesktopRootWindowHost because it is wrongly named, and if
named correctly it will conflict with the symbol from *X11.