[Android] Specify version code/name for android outputs at gyp time
The version in AndroidManifest.xml has two parts:versionCode and versionName.
VersionCode is an increasing integer to identifing upgrade.
VersionName is a human readable string.

For xwalk, VERSION file should be the only place to maintain version.
So this patch removes the hard coding in AndroidManifest.xml.

Insteadly, version code and version name will be specified as parameters at
gyp time for android's ant action.

For versionCode, it's calculated through version.
The formula is:
versionCode = shift +
              patch * 10 +
              build * 1000 +
              (major+minor) * 100000
Shift is the configuration bit. Currently, 1 for x86, 2 for armv7 and 3 for arm.