[Tools] Add increment-version.py.
It is a more generic version of bump-canary-version.sh, and allows one to
increment either the build version (for Crosswalk canaries) and the patch
version (for Crosswalk betas).

Additionally, the new tool is written in Python instead of a mixture of
shell script and Perl, so it also fixes #699.

bump-canary-version.sh will be removed in a follow-up commit because the
existing canary build scripts need to be adjusted to use the new script
after it is committed.

(cherry picked from commit 0f1297bf638f15c7d4af3fcbe24768024a49082b)