[Android] Handle exceptions during the cross package invoking
All exceptions was causing library-not-found dialog to show before.
It was not appropriate.

Now the implementation is that the exception are catagorized into:
1. Runtime library not installed;
2. Runtime library not up-to-date and it's critical;
3. Runtime library not up-to-date but bearable;
4. Error happening before version matching across package;
5. Error happening after version matching across package.
For the first four cases, a dialog will be shown to navigate user to
download/upgrade runtimelib with proper message.
For the last case, since the version is already matched, the exception
should be taken as a normal runtime exception and be thrown out.

BUG=https://github.com/crosswalk-project/crosswalk/issues/735